Fix virt-top --end-time option when TZ<>UTC (RHBZ#637964).
Return the end time as seconds since 1970, and don't adjust it
for timezone.  In the main module the end time is compared against
the result of Unix.gettimeofday, which is also seconds since 1970,
so the comparison now works.

Tested with TZ=UTC-05, TZ=UTC and TZ=UTC+05.

NB: This has not been tested against ocaml-calendar 1.x, although
that version of calendar is obsolete and no one should be using it.