Remove dependency on xml-light, replace with libxml2
This commit removes the dependency on the ancient and dead upstream
xml-light light.  It is replaced with the C library libxml2.  This
introduces a small amount of C code to interface with libxml2 and
parse the libvirt domain XML into the list of block devices and
network devices (using XPath).